split each class in a separate file, inheritance with base class
